requirements.txt
spond==1.1.1
requests==2.32.5
pytz==2025.2
pyyaml==6.0.3
